combined the three list command into one class, the junit testing and javadoc are not fixed yet
